# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS «Медицина Долголетия» — API
  version: "1.0.0"
  description: |
    API платформы обучения врачей. Авторизация — через httpOnly cookie `token` (JWT),
    выдаётся эндпоинтами `/api/auth/login` и `/api/auth/register`.
    Роли: `student` (врач), `curator`, `admin`, `super_admin`.
servers:
  - url: /api
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /course:
    get:
      tags: [Course]
      summary: Курс с уроками, тестом (без ответов) и прогрессом текущего врача
      responses: { "200": { description: OK }, "403": { description: "Только для роли student" } }
  /course/lesson-done:
    post:
      tags: [Course]
      summary: Отметить урок пройденным
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [lessonId], properties: { lessonId: { type: string } } }
      responses: { "200": { description: OK }, "403": { description: "Доступ ограничен или урок скрыт" } }
  /course/quiz-submit:
    post:
      tags: [Course]
      summary: Отправить ответы теста (балл считается на сервере)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                answers: { type: object, additionalProperties: { type: integer } }
      responses: { "200": { description: OK } }
...
  /course/certificate/{studentId}/issue:
    post:
      tags: [Course]
      summary: Выдать сертификат врачу (curator+)
      parameters: [{ name: studentId, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /course/visibility/{targetId}:
    put:
      tags: [Course]
      summary: "Задать список врачей, от которых скрыт урок/тест (targetId = id урока или \"quiz\")"
      parameters: [{ name: targetId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { ids: { type: array, items: { type: string } } } }
      responses: { "200": { description: OK } }
  /course/lessons/{id}:
    get:
      tags: [Course]
      summary: Полное содержимое урока (для редактора, admin/super_admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /course/lessons/{id}/draft:
    put:
      tags: [Course]
      summary: Сохранить черновик урока (не публикуется автоматически)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, html]
              properties:
                title: { type: string }
                duration: { type: string }
                html: { type: string }
      responses: { "200": { description: OK } }
  /course/lessons/{id}/publish:
    post:
      tags: [Course]
      summary: Опубликовать черновик (прежняя версия уходит в историю)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "400": { description: "Нет черновика" } }
  /course/lessons/{id}/history:
    get:
      tags: [Course]
      summary: История версий урока (последние 20)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /course/lessons/{id}/restore/{historyId}:
    post:
      tags: [Course]
      summary: Откатить урок к прежней версии из истории
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: historyId, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: OK } }
...
  /course/quiz-admin/{id}:
    put:
      tags: [Course]
      summary: Редактировать вопрос теста
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string }
                options: { type: array, items: { type: string } }
                correct: { type: integer }
      responses: { "200": { description: OK } }
...